+++ selfdrive/ui/replay/replay.h
@@ -8,6 +8,16 @@
 // one segment uses about 100M of memory
 constexpr int FORWARD_SEGS = 5;
 
+enum REPLAY_FLAGS {
+  REPLAY_FLAG_NONE = 0x0000,
+  REPLAY_FLAG_DCAM = 0x0002,
+  REPLAY_FLAG_ECAM = 0x0004,
+  REPLAY_FLAG_NO_LOOP = 0x0010,
+  REPLAY_FLAG_NO_FILE_CACHE = 0x0020,
+  REPLAY_FLAG_QCAMERA = 0x0040,
+  REPLAY_FLAG_SEND_YUV = 0x0080,
+};
+
 class Replay : public QObject {
   Q_OBJECT
